backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
launchdarkly-server-sdk>=9.12.0
launchdarkly-observability>=0.1.0
python-dotenv>=1.0.0
opentelemetry-instrumentation-flask>=0.48b0
opentelemetry-instrumentation-requests>=0.48b0
requests>=2.31.0


